Fix lint, small refactor
